Refactoring: move/delete several elements

+++ elements/brn/wifi/qos/tos2queuemapper.hh
@@ -146,7 +146,7 @@ public:
   BackoffScheme *_current_scheme;
   BackoffScheme **_bo_schemes;
 
-  uint32_t _no_schemes;
+  uint16_t _no_schemes;
   uint16_t _scheme_id;
 
 };
